corrected issue with saving empty password for wifi

+++ wap/pkg/wifi/wifi.go
@@ -399,32 +399,33 @@ func connectToFirstWifi(ctx context.Context, connections []string) error {
 		return err
 	}
 
-	// Create a temporary file
-	tempFile, err := os.CreateTemp("", "passwd-")
-	if err != nil {
-		return fmt.Errorf("failed to create temporary file: %w", err)
-	}
-	defer os.Remove(tempFile.Name())
+	if passwd != "" {
+		// Create a temporary file
+		tempFile, err := os.CreateTemp("", "passwd-")
+		if err != nil {
+			return fmt.Errorf("failed to create temporary file: %w", err)
+		}
+		defer os.Remove(tempFile.Name())
+		// Write the password to the temporary file
+		_, err = tempFile.WriteString(fmt.Sprintf("802-11-wireless-security.psk:%s", passwd))
+		if err != nil {
+			return fmt.Errorf("failed to write password to temporary file: %w", err)
+		}
+		err = tempFile.Close()
+		if err != nil {
+			return fmt.Errorf("failed to close temporary file: %w", err)
+		}
 
-	// Write the password to the temporary file
-	_, err = tempFile.WriteString(fmt.Sprintf("802-11-wireless-security.psk:%s", passwd))
-	if err != nil {
-		return fmt.Errorf("failed to write password to temporary file: %w", err)
-	}
-	err = tempFile.Close()
-	if err != nil {
-		return fmt.Errorf("failed to close temporary file: %w", err)
-	}
+		_, _, err = runCommand(ctx, fmt.Sprintf("nmcli connection up %s passwd-file %s", connection, tempFile.Name()))
 
-	_, _, err = runCommand(ctx, fmt.Sprintf("nmcli connection up %s passwd-file %s", connection, tempFile.Name()))
-
-	if err != nil {
-		log.Warnf("failed to connect to Wi-Fi: %v", err)
-		log.Info("Trying to recreate the connection profile")
-		DeleteConnection(ctx, connection)
-		createConnection(ctx, connection, connection, passwd)
-		err = connectToNetwork(ctx, connection)
-		return err
+		if err != nil {
+			log.Warnf("failed to connect to Wi-Fi: %v", err)
+			log.Info("Trying to recreate the connection profile")
+			DeleteConnection(ctx, connection)
+			createConnection(ctx, connection, connection, passwd)
+			err = connectToNetwork(ctx, connection)
+			return err
+		}
 	}
 	return nil
 }
